Pass settings via REST API
