+++ packages/editor/src/index.ts
@@ -2,12 +2,7 @@
  * Internal dependencies
  */
 
-// The __webpack_public_path__ assignment will be done after the imports.
-// That's why the public path assignment is in its own dedicated module and imported here at the very top.
-// See https://webpack.js.org/guides/public-path/#on-the-fly
-import './public-path';
-
-import './upload-queue';
+import './init';
 import './transforms';
 import './gif-block';
 import './cross-origin-isolation';
