--- packages/editor/src/public-path.ts
@@ -1,1 +1,0 @@
-__webpack_public_path__ = window.mediaExperiments.publicPath;
