+++ packages/view-upload-request/package.json
@@ -36,6 +36,7 @@
 		"@wordpress/components": "^28.0.0",
 		"@wordpress/compose": "^7.0.0",
 		"@wordpress/core-data": "^7.5.0",
-		"@wordpress/data": "^10.0.0"
+		"@wordpress/data": "^10.0.0",
+		"@wordpress/preferences": "^4.0.0"
 	}
 }
