+++ packages/view-upload-request/src/constants.ts
@@ -1,0 +1,1 @@
+export const PREFERENCES_NAME = 'media-experiments/preferences';
