+++ packages/view-upload-request/src/index.tsx
@@ -1,246 +1,112 @@
 /**
  * External dependencies
  */
-import type { ChangeEvent } from 'react';
+import { type ImageFormat, store as uploadStore } from '@mexp/upload-media';
 
 /**
  * WordPress dependencies
  */
-import {
-	StrictMode,
-	createRoot,
-	useEffect,
-	useRef,
-	useState,
-} from '@wordpress/element';
-import {
-	Button,
-	FormFileUpload,
-	Spinner,
-	SnackbarList,
-	// eslint-disable-next-line @wordpress/no-unsafe-wp-apis
-	__experimentalText as Text,
-} from '@wordpress/components';
-import { __ } from '@wordpress/i18n';
-import { dispatch, useDispatch, useSelect } from '@wordpress/data';
-import { store as noticesStore } from '@wordpress/notices';
-
-import {
-	type Attachment,
-	uploadMedia as originalUploadMedia,
-	sideloadMedia as originalSideloadMedia,
-	validateFileSize,
-	validateMimeType,
-} from '@mexp/media-utils';
-import { store as uploadStore } from '@mexp/upload-media';
+import { StrictMode, createRoot } from '@wordpress/element';
+import { dispatch, select, subscribe } from '@wordpress/data';
+import { store as preferencesStore } from '@wordpress/preferences';
 
 /**
  * Internal dependencies
  */
-import './view.css';
+import { store as coreStore } from '@wordpress/core-data';
+import { PREFERENCES_NAME } from './constants';
+import type { RestBaseRecord } from './types';
 
-/**
- * Upload a media file when the file upload button is activated.
- *
- * Similar to the mediaUpload() function from `@wordpress/editor`,
- * this is a wrapper around uploadMedia() from `@mexp/media-utils`.
- *
- * @param $0
- * @param $0.additionalData
- * @param $0.filesList
- * @param $0.onError
- * @param $0.onFileChange
- */
-function uploadMedia( {
-	additionalData = {},
-	filesList,
-	onError = noop,
-	onFileChange,
-}: Parameters< typeof originalUploadMedia >[ 0 ] ) {
-	originalUploadMedia( {
-		allowedTypes: window.mediaExperiments.allowedTypes,
-		wpAllowedMimeTypes: window.mediaExperiments.allowedMimeTypes,
-		maxUploadFileSize: window.mediaExperiments.maxUploadFileSize,
-		filesList,
-		additionalData,
-		onFileChange,
-		onError,
-	} );
+import { App } from './app';
+
+function getExtension( mimeType: string ): ImageFormat {
+	return ( mimeType.split( '/' )[ 1 ] || 'jpeg' ) as ImageFormat;
 }
 
-const noop = () => {};
+// Initialize default settings as soon as base data is available.
+const unsubscribeCoreStore = subscribe( () => {
+	const siteData = select( coreStore ).getEntityRecord<
+		// @ts-ignore
+		RestBaseRecord | undefined
+	>( 'root', '__unstableBase', undefined, {
+		_fields: [
+			'image_size_threshold',
+			'video_size_threshold',
+			'image_output_formats',
+			'jpeg_interlaced',
+			'png_interlaced',
+			'gif_interlaced',
+			'image_sizes',
+		],
+	} );
 
-/**
- * Upload a media file when the file upload button is activated.
- *
- * @param $0                   Parameters object passed to the function.
- * @param $0.allowedTypes      Array with the types of media that can be uploaded, if unset all types are allowed.
- * @param $0.additionalData    Additional data to include in the request.
- * @param $0.filesList         List of files.
- * @param $0.maxUploadFileSize Maximum upload size in bytes allowed for the site.
- * @param $0.onError           Function called when an error happens.
- * @param $0.onFileChange      Function called each time a file or a temporary representation of the file is available.
- */
-function uploadRequestUploadMedia( {
-	allowedTypes,
-	additionalData = {},
-	filesList,
-	maxUploadFileSize,
-	onError = noop,
-	onFileChange,
-}: Parameters< typeof originalUploadMedia >[ 0 ] ) {
-	const validFiles = [];
-
-	for ( const mediaFile of filesList ) {
-		// TODO: Consider using the _async_ isHeifImage() function from `@mexp/upload-media`
-		const isHeifImage = [ 'image/heic', 'image/heif' ].includes(
-			mediaFile.type
-		);
-
-		/*
-		 Check if the caller (e.g. a block) supports this mime type.
-		 Special case for file types such as HEIC which will be converted before upload anyway.
-		 Another check will be done before upload.
-		*/
-		if ( ! isHeifImage ) {
-			try {
-				validateMimeType( mediaFile, allowedTypes );
-			} catch ( error: unknown ) {
-				onError( error as Error );
-				continue;
-			}
-		}
-
-		// Verify if file is greater than the maximum file upload size allowed for the site.
-		// TODO: Consider removing, as file could potentially be compressed first.
-		try {
-			validateFileSize( mediaFile, maxUploadFileSize );
-		} catch ( error: unknown ) {
-			onError( error as Error );
-			continue;
-		}
-
-		validFiles.push( mediaFile );
+	if ( ! siteData ) {
+		return;
 	}
 
-	void dispatch( uploadStore ).addItems( {
-		files: validFiles,
-		onChange: onFileChange,
-		onError,
-		additionalData,
-	} );
-}
-
-/*
- Make the upload queue aware of the function for uploading to the server.
- The list of available image sizes is passed via an inline script
- and needs to be saved in the store first.
-*/
-void dispatch( uploadStore ).updateSettings( {
-	mediaUpload: uploadMedia,
-	mediaSideload: originalSideloadMedia,
-	imageSizes: window.mediaExperiments.availableImageSizes,
-} );
-
-function App() {
-	const { createErrorNotice, createSuccessNotice, removeNotice } =
-		useDispatch( noticesStore );
-
-	const isUploading = useSelect(
-		( select ) => select( uploadStore ).getItems().length > 0,
-		[]
-	);
-	const wasUploading = useRef( false );
-
-	useEffect( () => {
-		if ( ! isUploading && wasUploading.current ) {
-			void createSuccessNotice(
-				__(
-					'Media successfully uploaded. You may now close this page.',
-					'media-experiments'
-				),
-				{ type: 'snackbar' }
-			);
-		}
-
-		wasUploading.current = isUploading;
-	}, [ createSuccessNotice, isUploading ] );
-
-	const notices = useSelect(
-		( select ) => select( noticesStore ).getNotices(),
-		[]
-	);
-	const snackbarNotices = notices.filter(
-		( { type } ) => type === 'snackbar'
-	);
-
-	const [ attachment, setAttachment ] = useState< Partial< Attachment > >(
-		{}
-	);
-
-	const onUpload = ( event: ChangeEvent< HTMLInputElement > ) => {
-		uploadRequestUploadMedia( {
-			filesList: event.target.files ? [ ...event.target.files ] : [],
-			onError: ( error ) => {
-				void createErrorNotice( error.message, { type: 'snackbar' } );
-			},
-			additionalData: {
-				upload_request: window.mediaExperiments.uploadRequest,
-			},
-			onFileChange: ( [ media ] ) => {
-				if ( ! media.id ) {
-					return;
-				}
-
-				setAttachment( media );
-			},
-		} );
+	const defaultPreferences = {
+		// General.
+		requireApproval: true,
+		optimizeOnUpload: true,
+		thumbnailGeneration: 'smart',
+		imageLibrary: window.crossOriginIsolated ? 'vips' : 'browser',
+		bigImageSizeThreshold: siteData.image_size_threshold,
+		bigVideoSizeThreshold: siteData.video_size_threshold,
+		keepOriginal: false,
+		// Formats.
+		default_outputFormat: 'jpeg',
+		default_quality: 82,
+		default_interlaced: siteData.jpeg_interlaced,
+		jpeg_outputFormat: getExtension(
+			siteData.image_output_formats[ 'image/jpeg' ] || 'image/jpeg'
+		),
+		jpeg_quality: 82,
+		jpeg_interlaced: siteData.jpeg_interlaced,
+		png_outputFormat: getExtension(
+			siteData.image_output_formats[ 'image/png' ] || 'image/png'
+		),
+		png_quality: 82,
+		png_interlaced: siteData.png_interlaced,
+		webp_outputFormat: getExtension(
+			siteData.image_output_formats[ 'image/webp' ] || 'image/webp'
+		),
+		webp_quality: 86,
+		webp_interlaced: false,
+		avif_outputFormat: getExtension(
+			siteData.image_output_formats[ 'image/avif' ] || 'image/avif'
+		),
+		avif_quality: 80,
+		avif_interlaced: false,
+		heic_outputFormat: getExtension(
+			siteData.image_output_formats[ 'image/heic' ] || 'image/jpeg'
+		),
+		heic_quality: 80,
+		heic_interlaced: siteData.jpeg_interlaced,
+		gif_outputFormat: getExtension(
+			siteData.image_output_formats[ 'image/gif' ] || 'image/webp'
+		),
+		gif_quality: 80,
+		gif_interlaced: siteData.gif_interlaced,
+		gif_convert: true,
+		video_outputFormat: 'mp4',
+		audio_outputFormat: 'mp3',
+		// Media recording.
+		videoInput: undefined,
+		audioInput: undefined,
+		videoEffect: 'none',
 	};
 
-	return (
-		<>
-			{ ! attachment.id ? (
-				<FormFileUpload
-					onChange={ onUpload }
-					accept={
-						window.mediaExperiments.accept
-							? window.mediaExperiments.accept.join( ',' )
-							: '*'
-					}
-					multiple={ window.mediaExperiments.multiple }
-					render={ ( { openFileDialog } ) => (
-						<Button
-							variant="secondary"
-							onClick={ openFileDialog }
-							disabled={ isUploading }
-						>
-							{ isUploading ? (
-								<>
-									<Spinner />
-									{ __( 'Uploading…', 'media-experiments' ) }
-								</>
-							) : (
-								__( 'Upload media', 'media-experiments' )
-							) }
-						</Button>
-					) }
-				/>
-			) : (
-				<Text>
-					{ __(
-						'Media successfully uploaded. You may now close this page.',
-						'media-experiments'
-					) }
-				</Text>
-			) }
-			<SnackbarList
-				notices={ snackbarNotices }
-				className="components-editor-notices__snackbar"
-				onRemove={ removeNotice }
-			/>
-		</>
+	void dispatch( preferencesStore ).setDefaults(
+		PREFERENCES_NAME,
+		defaultPreferences
 	);
-}
+
+	void dispatch( uploadStore ).updateSettings( {
+		imageSizes: siteData.image_sizes,
+	} );
+
+	unsubscribeCoreStore();
+}, coreStore );
 
 const container = document.getElementById(
 	'media-experiments-upload-request-root'
